Make the addon entrypoint as smaller as possible
